Testing settings modal
Right now, most of the tests in home screen are based on redux and async
actions for initialization of state.
I'll go back to those failing tests just later, for now let's deal with
redux actions for the app slice to start and stop background music
as well as global toast notification.